delete TokenGenerator.java, reorganize
