--- board-server/src/main/java/net/stzups/board/data/TokenGenerator.java
@@ -1,9 +1,0 @@
-package net.stzups.board.data;
-
-import java.security.SecureRandom;
-import java.util.Random;
-
-public class TokenGenerator {
-
-
-}
